+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadChestBlock.java
@@ -105,9 +105,11 @@ public class KeypadChestBlock extends ChestBlock implements IOverlayDisplay, IDi
 			return Optional.empty();
 		}
 	};
+	private final float destroyTimeForOwner;
 
 	public KeypadChestBlock(BlockBehaviour.Properties properties) {
-		super(properties, SCContent.KEYPAD_CHEST_BLOCK_ENTITY);
+		super(OwnableBlock.withReinforcedDestroyTime(properties), SCContent.KEYPAD_CHEST_BLOCK_ENTITY);
+		destroyTimeForOwner = OwnableBlock.getStoredDestroyTime();
 	}
 
 	@Override
@@ -117,7 +119,7 @@ public class KeypadChestBlock extends ChestBlock implements IOverlayDisplay, IDi
 		if (disguisedState.getBlock() != state.getBlock())
 			return disguisedState.getDestroyProgress(player, level, pos);
 		else
-			return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
+			return BlockUtils.getDestroyProgress(super::getDestroyProgress, destroyTimeForOwner, state, player, level, pos);
 	}
 
 	@Override